+++ src/ploomber/__init__.py
@@ -7,7 +7,7 @@ from ploomber.env.decorators import load_env, with_env
 from ploomber.placeholders.SourceLoader import SourceLoader
 from ploomber.jupyter.manager import _load_jupyter_server_extension
 
-__version__ = '0.8.6dev'
+__version__ = '0.8.7dev'
 
 # Set default logging handler to avoid "No handler found" warnings.