+++ src/ploomber/cli/cli.py
@@ -1,17 +1,12 @@
 import os
-import re
 import sys
 from pathlib import Path
 
 import click
 from ploomber import __version__
 from ploomber import cli as cli_module
-from ploomber import scaffold
-
-
-def _is_valid_name(package_name):
-    match = re.match(r'^[\w-]+$', package_name) or False
-    return match and not package_name[0].isnumeric()
+from ploomber import scaffold as _scaffold
+from ploomber_scaffold import scaffold as scaffold_project
 
 
 @click.group()
@@ -23,75 +18,13 @@ def cli():
 
 
 @cli.command()
-def new():
-    """Create a new project
+def scaffold():
+    """Create new projects and add template tasks
     """
-    _new()
-
-
-@cli.command()
-def add():
-    """Create source files tasks registered in pipeline.yaml
-    """
-    scaffold.add()
-
-
-def _new():
-    click.echo('This utility will guide you through the process of creating '
-               'a new project')
-
-    valid_name = False
-
-    while not valid_name:
-        name = click.prompt(
-            'Enter a name for your project (allowed: '
-            'alphanumeric, underscores and hyphens)',
-            type=str)
-        valid_name = _is_valid_name(name)
-
-        if not valid_name:
-            click.echo('"%s" is not a valid project name, choose another.' %
-                       name)
-
-    env = scaffold.ScaffoldLoader('ploomber_new', project_name=name)
-
-    click.echo('Creating %s/' % name)
-    Path(name).mkdir()
-
-    db = click.confirm('Do you need to connect to a database?')
-
-    if db:
-        click.echo('Adding db.py...')
-        env.copy('db.py')
-
-    click.echo('Adding pipeline.yaml...')
-    content = env.get_template('pipeline.yaml').render(db=db)
-    Path(name, 'pipeline.yaml').write_text(content)
-
-    conda = click.confirm('Do you you want to use conda to '
-                          'manage virtual environments (recommended)?')
-
-    if conda:
-        # check if conda is installed...
-        click.echo('Adding environment.yml...')
-        content = env.get_template('environment.yml').render(name=name)
-        Path(name, 'environment.yml').write_text(content)
-
-    click.echo('Adding README.md...')
-    content = env.get_template('README.md').render(name=name,
-                                                   db=db,
-                                                   conda=conda)
-    path_to_readme = Path(name, 'README.md')
-    path_to_readme.write_text(content)
-
-    click.echo('Adding raw.py, clean.py and plot.py...')
-    env.copy('raw.py')
-    env.copy('clean.py')
-    env.copy('plot.py')
-    click.echo('Done!')
-
-    Path(name, 'output').mkdir()
-    click.echo('Check out {} to get started'.format(path_to_readme))
+    if Path('pipeline.yaml').exists():
+        _scaffold.add()
+    else:
+        scaffold_project.cli(project_path=None)
 
 
 def cmd_router():