+++ tests/test_scaffold.py
@@ -10,7 +10,7 @@ from ploomber import scaffold
 @pytest.mark.parametrize('name', ['task.py', 'task.ipynb'])
 @pytest.mark.parametrize('extract_upstream', [False, True])
 @pytest.mark.parametrize('extract_product', [False, True])
-def test_render_script(name, extract_product, extract_upstream):
+def test_renders_valid_script(name, extract_product, extract_upstream):
     loader = scaffold.ScaffoldLoader('ploomber_add')
     out = loader.render(name,
                         params=dict(extract_product=extract_product,
@@ -23,7 +23,7 @@ def test_render_script(name, extract_product, extract_upstream):
 
 @pytest.mark.parametrize('extract_upstream', [False, True])
 @pytest.mark.parametrize('extract_product', [False, True])
-def test_render_function(extract_product, extract_upstream):
+def test_renders_valid_function(extract_product, extract_upstream):
     loader = scaffold.ScaffoldLoader('ploomber_add')
     out = loader.render('function.py',
                         params=dict(function_name='some_function',
@@ -52,7 +52,7 @@ def test_create_function(backup_test_pkg, tmp_directory):
     assert 'new_function' in function_names
 
 
-def test_add_scaffold(backup_test_pkg, tmp_directory):
+def test_add_task_from_scaffold(backup_test_pkg, tmp_directory):
     yaml = """
     meta:
         source_loader: