+++ CHANGELOG.rst
@@ -1,8 +1,15 @@
 CHANGELOG
 =========
 
-0.8.6dev
+0.8.7dev
 --------
+* Deprecates ``ploomber new`` and ``ploomber add``
+* Adds ``ploomber scaffold``
+* Jupyter plugin now exports functions as notebooks using ``jupyter_functions_as_notebooks`` in ``pipeline.yaml``
+
+
+0.8.6 (2021-01-08)
+-------------------
 * "ploomber add" generates template tasks and functions if they don't exist
 * Jupyter plugin now shows PythonCallable tasks as notebooks